requirements.txt
# Core ML & Data Processing
pyspark==3.5.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
mlflow>=2.9.0
optuna>=3.5.0

# API & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
streamlit>=1.29.0
plotly>=6.0.0

# Streaming & Cache
kafka-python>=2.0.0
redis>=5.0.0
py4j>=0.10.9

# Monitoring & Testing
prometheus-client>=0.19.0
pytest>=7.4.0
pytest-cov>=4.1.0
locust>=2.20.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
httpx>=0.26.0
